requirements.txt
streamlit==0.79.0
pandas==1.1.5
matplotlib.pyplot==1.1.5
seaborn==0.11.1
joblib==1.0.1